+++ blockbase/inc/wp-customize-colors.php
@@ -135,6 +135,3 @@ class GlobalStylesCustomizer {
 }
 
 new GlobalStylesCustomizer;
-
-//this is temporary so it shows after the color customizations:
-require_once 'wp-customize-palette-control.php';
